+++ google/pubsub_v1/types/pubsub/__init__.py
@@ -73,6 +73,46 @@ __protobuf__ = proto.module(
 )
 
 
+from .requests import (
+    GetTopicRequest,
+    UpdateTopicRequest,
+    PublishRequest,
+    ListTopicsRequest,
+    ListTopicSubscriptionsRequest,
+    ListTopicSnapshotsRequest,
+    DeleteTopicRequest,
+    DetachSubscriptionRequest,
+    GetSubscriptionRequest,
+    UpdateSubscriptionRequest,
+    ListSubscriptionsRequest,
+    DeleteSubscriptionRequest,
+    ModifyPushConfigRequest,
+    PullRequest,
+    ModifyAckDeadlineRequest,
+    AcknowledgeRequest,
+    StreamingPullRequest,
+    CreateSnapshotRequest,
+    UpdateSnapshotRequest,
+    GetSnapshotRequest,
+    ListSnapshotsRequest,
+    DeleteSnapshotRequest,
+    SeekRequest,
+)
+
+from .responses import (
+    PublishResponse,
+    ListTopicsResponse,
+    ListTopicSubscriptionsResponse,
+    ListTopicSnapshotsResponse,
+    DetachSubscriptionResponse,
+    ListSubscriptionsResponse,
+    PullResponse,
+    StreamingPullResponse,
+    ListSnapshotsResponse,
+    SeekResponse,
+)
+
+
 class MessageStoragePolicy(proto.Message):
     r"""A policy constraining the storage of messages published to
     the topic.
@@ -267,297 +307,6 @@ class PubsubMessage(proto.Message):
         proto.STRING,
         number=5,
     )
-
-
-class GetTopicRequest(proto.Message):
-    r"""Request for the GetTopic method.
-
-    Attributes:
-        topic (str):
-            Required. The name of the topic to get. Format is
-            ``projects/{project}/topics/{topic}``.
-    """
-
-    topic = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class UpdateTopicRequest(proto.Message):
-    r"""Request for the UpdateTopic method.
-
-    Attributes:
-        topic (google.pubsub_v1.types.Topic):
-            Required. The updated topic object.
-        update_mask (google.protobuf.field_mask_pb2.FieldMask):
-            Required. Indicates which fields in the provided topic to
-            update. Must be specified and non-empty. Note that if
-            ``update_mask`` contains "message_storage_policy" but the
-            ``message_storage_policy`` is not set in the ``topic``
-            provided above, then the updated value is determined by the
-            policy configured at the project or organization level.
-    """
-
-    topic = proto.Field(
-        proto.MESSAGE,
-        number=1,
-        message="Topic",
-    )
-    update_mask = proto.Field(
-        proto.MESSAGE,
-        number=2,
-        message=field_mask_pb2.FieldMask,
-    )
-
-
-class PublishRequest(proto.Message):
-    r"""Request for the Publish method.
-
-    Attributes:
-        topic (str):
-            Required. The messages in the request will be published on
-            this topic. Format is ``projects/{project}/topics/{topic}``.
-        messages (Sequence[google.pubsub_v1.types.PubsubMessage]):
-            Required. The messages to publish.
-    """
-
-    topic = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    messages = proto.RepeatedField(
-        proto.MESSAGE,
-        number=2,
-        message="PubsubMessage",
-    )
-
-
-class PublishResponse(proto.Message):
-    r"""Response for the ``Publish`` method.
-
-    Attributes:
-        message_ids (Sequence[str]):
-            The server-assigned ID of each published
-            message, in the same order as the messages in
-            the request. IDs are guaranteed to be unique
-            within the topic.
-    """
-
-    message_ids = proto.RepeatedField(
-        proto.STRING,
-        number=1,
-    )
-
-
-class ListTopicsRequest(proto.Message):
-    r"""Request for the ``ListTopics`` method.
-
-    Attributes:
-        project (str):
-            Required. The name of the project in which to list topics.
-            Format is ``projects/{project-id}``.
-        page_size (int):
-            Maximum number of topics to return.
-        page_token (str):
-            The value returned by the last ``ListTopicsResponse``;
-            indicates that this is a continuation of a prior
-            ``ListTopics`` call, and that the system should return the
-            next page of data.
-    """
-
-    project = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    page_size = proto.Field(
-        proto.INT32,
-        number=2,
-    )
-    page_token = proto.Field(
-        proto.STRING,
-        number=3,
-    )
-
-
-class ListTopicsResponse(proto.Message):
-    r"""Response for the ``ListTopics`` method.
-
-    Attributes:
-        topics (Sequence[google.pubsub_v1.types.Topic]):
-            The resulting topics.
-        next_page_token (str):
-            If not empty, indicates that there may be more topics that
-            match the request; this value should be passed in a new
-            ``ListTopicsRequest``.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    topics = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message="Topic",
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-class ListTopicSubscriptionsRequest(proto.Message):
-    r"""Request for the ``ListTopicSubscriptions`` method.
-
-    Attributes:
-        topic (str):
-            Required. The name of the topic that subscriptions are
-            attached to. Format is
-            ``projects/{project}/topics/{topic}``.
-        page_size (int):
-            Maximum number of subscription names to
-            return.
-        page_token (str):
-            The value returned by the last
-            ``ListTopicSubscriptionsResponse``; indicates that this is a
-            continuation of a prior ``ListTopicSubscriptions`` call, and
-            that the system should return the next page of data.
-    """
-
-    topic = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    page_size = proto.Field(
-        proto.INT32,
-        number=2,
-    )
-    page_token = proto.Field(
-        proto.STRING,
-        number=3,
-    )
-
-
-class ListTopicSubscriptionsResponse(proto.Message):
-    r"""Response for the ``ListTopicSubscriptions`` method.
-
-    Attributes:
-        subscriptions (Sequence[str]):
-            The names of subscriptions attached to the
-            topic specified in the request.
-        next_page_token (str):
-            If not empty, indicates that there may be more subscriptions
-            that match the request; this value should be passed in a new
-            ``ListTopicSubscriptionsRequest`` to get more subscriptions.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    subscriptions = proto.RepeatedField(
-        proto.STRING,
-        number=1,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-class ListTopicSnapshotsRequest(proto.Message):
-    r"""Request for the ``ListTopicSnapshots`` method.
-
-    Attributes:
-        topic (str):
-            Required. The name of the topic that snapshots are attached
-            to. Format is ``projects/{project}/topics/{topic}``.
-        page_size (int):
-            Maximum number of snapshot names to return.
-        page_token (str):
-            The value returned by the last
-            ``ListTopicSnapshotsResponse``; indicates that this is a
-            continuation of a prior ``ListTopicSnapshots`` call, and
-            that the system should return the next page of data.
-    """
-
-    topic = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    page_size = proto.Field(
-        proto.INT32,
-        number=2,
-    )
-    page_token = proto.Field(
-        proto.STRING,
-        number=3,
-    )
-
-
-class ListTopicSnapshotsResponse(proto.Message):
-    r"""Response for the ``ListTopicSnapshots`` method.
-
-    Attributes:
-        snapshots (Sequence[str]):
-            The names of the snapshots that match the
-            request.
-        next_page_token (str):
-            If not empty, indicates that there may be more snapshots
-            that match the request; this value should be passed in a new
-            ``ListTopicSnapshotsRequest`` to get more snapshots.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    snapshots = proto.RepeatedField(
-        proto.STRING,
-        number=1,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-class DeleteTopicRequest(proto.Message):
-    r"""Request for the ``DeleteTopic`` method.
-
-    Attributes:
-        topic (str):
-            Required. Name of the topic to delete. Format is
-            ``projects/{project}/topics/{topic}``.
-    """
-
-    topic = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class DetachSubscriptionRequest(proto.Message):
-    r"""Request for the DetachSubscription method.
-
-    Attributes:
-        subscription (str):
-            Required. The subscription to detach. Format is
-            ``projects/{project}/subscriptions/{subscription}``.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class DetachSubscriptionResponse(proto.Message):
-    r"""Response for the DetachSubscription method.
-    Reserved for future use.
-
-    """
 
 
 class Subscription(proto.Message):
@@ -1114,559 +863,6 @@ class ReceivedMessage(proto.Message):
     )
 
 
-class GetSubscriptionRequest(proto.Message):
-    r"""Request for the GetSubscription method.
-
-    Attributes:
-        subscription (str):
-            Required. The name of the subscription to get. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class UpdateSubscriptionRequest(proto.Message):
-    r"""Request for the UpdateSubscription method.
-
-    Attributes:
-        subscription (google.pubsub_v1.types.Subscription):
-            Required. The updated subscription object.
-        update_mask (google.protobuf.field_mask_pb2.FieldMask):
-            Required. Indicates which fields in the
-            provided subscription to update. Must be
-            specified and non-empty.
-    """
-
-    subscription = proto.Field(
-        proto.MESSAGE,
-        number=1,
-        message="Subscription",
-    )
-    update_mask = proto.Field(
-        proto.MESSAGE,
-        number=2,
-        message=field_mask_pb2.FieldMask,
-    )
-
-
-class ListSubscriptionsRequest(proto.Message):
-    r"""Request for the ``ListSubscriptions`` method.
-
-    Attributes:
-        project (str):
-            Required. The name of the project in which to list
-            subscriptions. Format is ``projects/{project-id}``.
-        page_size (int):
-            Maximum number of subscriptions to return.
-        page_token (str):
-            The value returned by the last
-            ``ListSubscriptionsResponse``; indicates that this is a
-            continuation of a prior ``ListSubscriptions`` call, and that
-            the system should return the next page of data.
-    """
-
-    project = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    page_size = proto.Field(
-        proto.INT32,
-        number=2,
-    )
-    page_token = proto.Field(
-        proto.STRING,
-        number=3,
-    )
-
-
-class ListSubscriptionsResponse(proto.Message):
-    r"""Response for the ``ListSubscriptions`` method.
-
-    Attributes:
-        subscriptions (Sequence[google.pubsub_v1.types.Subscription]):
-            The subscriptions that match the request.
-        next_page_token (str):
-            If not empty, indicates that there may be more subscriptions
-            that match the request; this value should be passed in a new
-            ``ListSubscriptionsRequest`` to get more subscriptions.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    subscriptions = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message="Subscription",
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-class DeleteSubscriptionRequest(proto.Message):
-    r"""Request for the DeleteSubscription method.
-
-    Attributes:
-        subscription (str):
-            Required. The subscription to delete. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class ModifyPushConfigRequest(proto.Message):
-    r"""Request for the ModifyPushConfig method.
-
-    Attributes:
-        subscription (str):
-            Required. The name of the subscription. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-        push_config (google.pubsub_v1.types.PushConfig):
-            Required. The push configuration for future deliveries.
-
-            An empty ``pushConfig`` indicates that the Pub/Sub system
-            should stop pushing messages from the given subscription and
-            allow messages to be pulled and acknowledged - effectively
-            pausing the subscription if ``Pull`` or ``StreamingPull`` is
-            not called.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    push_config = proto.Field(
-        proto.MESSAGE,
-        number=2,
-        message="PushConfig",
-    )
-
-
-class PullRequest(proto.Message):
-    r"""Request for the ``Pull`` method.
-
-    Attributes:
-        subscription (str):
-            Required. The subscription from which messages should be
-            pulled. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-        return_immediately (bool):
-            Optional. If this field set to true, the system will respond
-            immediately even if it there are no messages available to
-            return in the ``Pull`` response. Otherwise, the system may
-            wait (for a bounded amount of time) until at least one
-            message is available, rather than returning no messages.
-            Warning: setting this field to ``true`` is discouraged
-            because it adversely impacts the performance of ``Pull``
-            operations. We recommend that users do not set this field.
-        max_messages (int):
-            Required. The maximum number of messages to
-            return for this request. Must be a positive
-            integer. The Pub/Sub system may return fewer
-            than the number specified.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    return_immediately = proto.Field(
-        proto.BOOL,
-        number=2,
-    )
-    max_messages = proto.Field(
-        proto.INT32,
-        number=3,
-    )
-
-
-class PullResponse(proto.Message):
-    r"""Response for the ``Pull`` method.
-
-    Attributes:
-        received_messages (Sequence[google.pubsub_v1.types.ReceivedMessage]):
-            Received Pub/Sub messages. The list will be empty if there
-            are no more messages available in the backlog. For JSON, the
-            response can be entirely empty. The Pub/Sub system may
-            return fewer than the ``maxMessages`` requested even if
-            there are more messages available in the backlog.
-    """
-
-    received_messages = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message="ReceivedMessage",
-    )
-
-
-class ModifyAckDeadlineRequest(proto.Message):
-    r"""Request for the ModifyAckDeadline method.
-
-    Attributes:
-        subscription (str):
-            Required. The name of the subscription. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-        ack_ids (Sequence[str]):
-            Required. List of acknowledgment IDs.
-        ack_deadline_seconds (int):
-            Required. The new ack deadline with respect to the time this
-            request was sent to the Pub/Sub system. For example, if the
-            value is 10, the new ack deadline will expire 10 seconds
-            after the ``ModifyAckDeadline`` call was made. Specifying
-            zero might immediately make the message available for
-            delivery to another subscriber client. This typically
-            results in an increase in the rate of message redeliveries
-            (that is, duplicates). The minimum deadline you can specify
-            is 0 seconds. The maximum deadline you can specify is 600
-            seconds (10 minutes).
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    ack_ids = proto.RepeatedField(
-        proto.STRING,
-        number=4,
-    )
-    ack_deadline_seconds = proto.Field(
-        proto.INT32,
-        number=3,
-    )
-
-
-class AcknowledgeRequest(proto.Message):
-    r"""Request for the Acknowledge method.
-
-    Attributes:
-        subscription (str):
-            Required. The subscription whose message is being
-            acknowledged. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-        ack_ids (Sequence[str]):
-            Required. The acknowledgment ID for the messages being
-            acknowledged that was returned by the Pub/Sub system in the
-            ``Pull`` response. Must not be empty.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    ack_ids = proto.RepeatedField(
-        proto.STRING,
-        number=2,
-    )
-
-
-class StreamingPullRequest(proto.Message):
-    r"""Request for the ``StreamingPull`` streaming RPC method. This request
-    is used to establish the initial stream as well as to stream
-    acknowledgements and ack deadline modifications from the client to
-    the server.
-
-    Attributes:
-        subscription (str):
-            Required. The subscription for which to initialize the new
-            stream. This must be provided in the first request on the
-            stream, and must not be set in subsequent requests from
-            client to server. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-        ack_ids (Sequence[str]):
-            List of acknowledgement IDs for acknowledging previously
-            received messages (received on this stream or a different
-            stream). If an ack ID has expired, the corresponding message
-            may be redelivered later. Acknowledging a message more than
-            once will not result in an error. If the acknowledgement ID
-            is malformed, the stream will be aborted with status
-            ``INVALID_ARGUMENT``.
-        modify_deadline_seconds (Sequence[int]):
-            The list of new ack deadlines for the IDs listed in
-            ``modify_deadline_ack_ids``. The size of this list must be
-            the same as the size of ``modify_deadline_ack_ids``. If it
-            differs the stream will be aborted with
-            ``INVALID_ARGUMENT``. Each element in this list is applied
-            to the element in the same position in
-            ``modify_deadline_ack_ids``. The new ack deadline is with
-            respect to the time this request was sent to the Pub/Sub
-            system. Must be >= 0. For example, if the value is 10, the
-            new ack deadline will expire 10 seconds after this request
-            is received. If the value is 0, the message is immediately
-            made available for another streaming or non-streaming pull
-            request. If the value is < 0 (an error), the stream will be
-            aborted with status ``INVALID_ARGUMENT``.
-        modify_deadline_ack_ids (Sequence[str]):
-            List of acknowledgement IDs whose deadline will be modified
-            based on the corresponding element in
-            ``modify_deadline_seconds``. This field can be used to
-            indicate that more time is needed to process a message by
-            the subscriber, or to make the message available for
-            redelivery if the processing was interrupted.
-        stream_ack_deadline_seconds (int):
-            Required. The ack deadline to use for the
-            stream. This must be provided in the first
-            request on the stream, but it can also be
-            updated on subsequent requests from client to
-            server. The minimum deadline you can specify is
-            10 seconds. The maximum deadline you can specify
-            is 600 seconds (10 minutes).
-        client_id (str):
-            A unique identifier that is used to distinguish client
-            instances from each other. Only needs to be provided on the
-            initial request. When a stream disconnects and reconnects
-            for the same stream, the client_id should be set to the same
-            value so that state associated with the old stream can be
-            transferred to the new stream. The same client_id should not
-            be used for different client instances.
-        max_outstanding_messages (int):
-            Flow control settings for the maximum number of outstanding
-            messages. When there are ``max_outstanding_messages`` or
-            more currently sent to the streaming pull client that have
-            not yet been acked or nacked, the server stops sending more
-            messages. The sending of messages resumes once the number of
-            outstanding messages is less than this value. If the value
-            is <= 0, there is no limit to the number of outstanding
-            messages. This property can only be set on the initial
-            StreamingPullRequest. If it is set on a subsequent request,
-            the stream will be aborted with status ``INVALID_ARGUMENT``.
-        max_outstanding_bytes (int):
-            Flow control settings for the maximum number of outstanding
-            bytes. When there are ``max_outstanding_bytes`` or more
-            worth of messages currently sent to the streaming pull
-            client that have not yet been acked or nacked, the server
-            will stop sending more messages. The sending of messages
-            resumes once the number of outstanding bytes is less than
-            this value. If the value is <= 0, there is no limit to the
-            number of outstanding bytes. This property can only be set
-            on the initial StreamingPullRequest. If it is set on a
-            subsequent request, the stream will be aborted with status
-            ``INVALID_ARGUMENT``.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    ack_ids = proto.RepeatedField(
-        proto.STRING,
-        number=2,
-    )
-    modify_deadline_seconds = proto.RepeatedField(
-        proto.INT32,
-        number=3,
-    )
-    modify_deadline_ack_ids = proto.RepeatedField(
-        proto.STRING,
-        number=4,
-    )
-    stream_ack_deadline_seconds = proto.Field(
-        proto.INT32,
-        number=5,
-    )
-    client_id = proto.Field(
-        proto.STRING,
-        number=6,
-    )
-    max_outstanding_messages = proto.Field(
-        proto.INT64,
-        number=7,
-    )
-    max_outstanding_bytes = proto.Field(
-        proto.INT64,
-        number=8,
-    )
-
-
-class StreamingPullResponse(proto.Message):
-    r"""Response for the ``StreamingPull`` method. This response is used to
-    stream messages from the server to the client.
-
-    Attributes:
-        received_messages (Sequence[google.pubsub_v1.types.ReceivedMessage]):
-            Received Pub/Sub messages. This will not be
-            empty.
-        acknowledge_confirmation (google.pubsub_v1.types.StreamingPullResponse.AcknowledgeConfirmation):
-            This field will only be set if
-            ``enable_exactly_once_delivery`` is set to ``true``.
-        modify_ack_deadline_confirmation (google.pubsub_v1.types.StreamingPullResponse.ModifyAckDeadlineConfirmation):
-            This field will only be set if
-            ``enable_exactly_once_delivery`` is set to ``true``.
-        subscription_properties (google.pubsub_v1.types.StreamingPullResponse.SubscriptionProperties):
-            Properties associated with this subscription.
-    """
-
-    class AcknowledgeConfirmation(proto.Message):
-        r"""Acknowledgement IDs sent in one or more previous requests to
-        acknowledge a previously received message.
-
-        Attributes:
-            ack_ids (Sequence[str]):
-                Successfully processed acknowledgement IDs.
-            invalid_ack_ids (Sequence[str]):
-                List of acknowledgement IDs that were
-                malformed or whose acknowledgement deadline has
-                expired.
-            unordered_ack_ids (Sequence[str]):
-                List of acknowledgement IDs that were out of
-                order.
-        """
-
-        ack_ids = proto.RepeatedField(
-            proto.STRING,
-            number=1,
-        )
-        invalid_ack_ids = proto.RepeatedField(
-            proto.STRING,
-            number=2,
-        )
-        unordered_ack_ids = proto.RepeatedField(
-            proto.STRING,
-            number=3,
-        )
-
-    class ModifyAckDeadlineConfirmation(proto.Message):
-        r"""Acknowledgement IDs sent in one or more previous requests to
-        modify the deadline for a specific message.
-
-        Attributes:
-            ack_ids (Sequence[str]):
-                Successfully processed acknowledgement IDs.
-            invalid_ack_ids (Sequence[str]):
-                List of acknowledgement IDs that were
-                malformed or whose acknowledgement deadline has
-                expired.
-        """
-
-        ack_ids = proto.RepeatedField(
-            proto.STRING,
-            number=1,
-        )
-        invalid_ack_ids = proto.RepeatedField(
-            proto.STRING,
-            number=2,
-        )
-
-    class SubscriptionProperties(proto.Message):
-        r"""Subscription properties sent as part of the response.
-
-        Attributes:
-            exactly_once_delivery_enabled (bool):
-                True iff exactly once delivery is enabled for
-                this subscription.
-            message_ordering_enabled (bool):
-                True iff message ordering is enabled for this
-                subscription.
-        """
-
-        exactly_once_delivery_enabled = proto.Field(
-            proto.BOOL,
-            number=1,
-        )
-        message_ordering_enabled = proto.Field(
-            proto.BOOL,
-            number=2,
-        )
-
-    received_messages = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message="ReceivedMessage",
-    )
-    acknowledge_confirmation = proto.Field(
-        proto.MESSAGE,
-        number=5,
-        message=AcknowledgeConfirmation,
-    )
-    modify_ack_deadline_confirmation = proto.Field(
-        proto.MESSAGE,
-        number=3,
-        message=ModifyAckDeadlineConfirmation,
-    )
-    subscription_properties = proto.Field(
-        proto.MESSAGE,
-        number=4,
-        message=SubscriptionProperties,
-    )
-
-
-class CreateSnapshotRequest(proto.Message):
-    r"""Request for the ``CreateSnapshot`` method.
-
-    Attributes:
-        name (str):
-            Required. User-provided name for this snapshot. If the name
-            is not provided in the request, the server will assign a
-            random name for this snapshot on the same project as the
-            subscription. Note that for REST API requests, you must
-            specify a name. See the resource name rules. Format is
-            ``projects/{project}/snapshots/{snap}``.
-        subscription (str):
-            Required. The subscription whose backlog the snapshot
-            retains. Specifically, the created snapshot is guaranteed to
-            retain: (a) The existing backlog on the subscription. More
-            precisely, this is defined as the messages in the
-            subscription's backlog that are unacknowledged upon the
-            successful completion of the ``CreateSnapshot`` request; as
-            well as: (b) Any messages published to the subscription's
-            topic following the successful completion of the
-            CreateSnapshot request. Format is
-            ``projects/{project}/subscriptions/{sub}``.
-        labels (Mapping[str, str]):
-            See <a
-            href="https://cloud.google.com/pubsub/docs/labels">
-            Creating and managing labels</a>.
-    """
-
-    name = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    subscription = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-    labels = proto.MapField(
-        proto.STRING,
-        proto.STRING,
-        number=3,
-    )
-
-
-class UpdateSnapshotRequest(proto.Message):
-    r"""Request for the UpdateSnapshot method.
-
-    Attributes:
-        snapshot (google.pubsub_v1.types.Snapshot):
-            Required. The updated snapshot object.
-        update_mask (google.protobuf.field_mask_pb2.FieldMask):
-            Required. Indicates which fields in the
-            provided snapshot to update. Must be specified
-            and non-empty.
-    """
-
-    snapshot = proto.Field(
-        proto.MESSAGE,
-        number=1,
-        message="Snapshot",
-    )
-    update_mask = proto.Field(
-        proto.MESSAGE,
-        number=2,
-        message=field_mask_pb2.FieldMask,
-    )
-
-
 class Snapshot(proto.Message):
     r"""A snapshot resource. Snapshots are used in
     `Seek <https://cloud.google.com/pubsub/docs/replay-overview>`__
@@ -1717,151 +913,6 @@ class Snapshot(proto.Message):
         proto.STRING,
         number=4,
     )
-
-
-class GetSnapshotRequest(proto.Message):
-    r"""Request for the GetSnapshot method.
-
-    Attributes:
-        snapshot (str):
-            Required. The name of the snapshot to get. Format is
-            ``projects/{project}/snapshots/{snap}``.
-    """
-
-    snapshot = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class ListSnapshotsRequest(proto.Message):
-    r"""Request for the ``ListSnapshots`` method.
-
-    Attributes:
-        project (str):
-            Required. The name of the project in which to list
-            snapshots. Format is ``projects/{project-id}``.
-        page_size (int):
-            Maximum number of snapshots to return.
-        page_token (str):
-            The value returned by the last ``ListSnapshotsResponse``;
-            indicates that this is a continuation of a prior
-            ``ListSnapshots`` call, and that the system should return
-            the next page of data.
-    """
-
-    project = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    page_size = proto.Field(
-        proto.INT32,
-        number=2,
-    )
-    page_token = proto.Field(
-        proto.STRING,
-        number=3,
-    )
-
-
-class ListSnapshotsResponse(proto.Message):
-    r"""Response for the ``ListSnapshots`` method.
-
-    Attributes:
-        snapshots (Sequence[google.pubsub_v1.types.Snapshot]):
-            The resulting snapshots.
-        next_page_token (str):
-            If not empty, indicates that there may be more snapshot that
-            match the request; this value should be passed in a new
-            ``ListSnapshotsRequest``.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    snapshots = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message="Snapshot",
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-class DeleteSnapshotRequest(proto.Message):
-    r"""Request for the ``DeleteSnapshot`` method.
-
-    Attributes:
-        snapshot (str):
-            Required. The name of the snapshot to delete. Format is
-            ``projects/{project}/snapshots/{snap}``.
-    """
-
-    snapshot = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-
-
-class SeekRequest(proto.Message):
-    r"""Request for the ``Seek`` method.
-
-    This message has `oneof`_ fields (mutually exclusive fields).
-    For each oneof, at most one member field can be set at the same time.
-    Setting any member of the oneof automatically clears all other
-    members.
-
-    .. _oneof: https://proto-plus-python.readthedocs.io/en/stable/fields.html#oneofs-mutually-exclusive-fields
-
-    Attributes:
-        subscription (str):
-            Required. The subscription to affect.
-        time (google.protobuf.timestamp_pb2.Timestamp):
-            The time to seek to. Messages retained in the subscription
-            that were published before this time are marked as
-            acknowledged, and messages retained in the subscription that
-            were published after this time are marked as unacknowledged.
-            Note that this operation affects only those messages
-            retained in the subscription (configured by the combination
-            of ``message_retention_duration`` and
-            ``retain_acked_messages``). For example, if ``time``
-            corresponds to a point before the message retention window
-            (or to a point before the system's notion of the
-            subscription creation time), only retained messages will be
-            marked as unacknowledged, and already-expunged messages will
-            not be restored.
-
-            This field is a member of `oneof`_ ``target``.
-        snapshot (str):
-            The snapshot to seek to. The snapshot's topic must be the
-            same as that of the provided subscription. Format is
-            ``projects/{project}/snapshots/{snap}``.
-
-            This field is a member of `oneof`_ ``target``.
-    """
-
-    subscription = proto.Field(
-        proto.STRING,
-        number=1,
-    )
-    time = proto.Field(
-        proto.MESSAGE,
-        number=2,
-        oneof="target",
-        message=timestamp_pb2.Timestamp,
-    )
-    snapshot = proto.Field(
-        proto.STRING,
-        number=3,
-        oneof="target",
-    )
-
-
-class SeekResponse(proto.Message):
-    r"""Response for the ``Seek`` method (this response is empty)."""
 
 
 __all__ = tuple(sorted(__protobuf__.manifest))
